DumBass_Records/requirements.txt
Flask
Faker
python3-discogs-client
mariadb
pymysql
pymongo
